@@ -9,17 +9,15 @@
     "GPL-3.0-or-later"
   ],
   "require": {
-    "php": ">=5.6.0",
-    "flow/jsonpath": "0.3.*",
-    "guzzlehttp/guzzle-services": "0.5.*",
-    "jms/serializer": "1.*",
-    "symfony/event-dispatcher": "^2.7",
-    "symfony/config": "^2.7",
-    "yosymfony/config-loader": "^1.3"
+    "php": "^7.1.3",
+    "flow/jsonpath": "^0.4.0",
+    "guzzlehttp/guzzle-services": "^1.1",
+    "jms/serializer": "^1.13",
+    "symfony/event-dispatcher": "^4.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "^5.5",
-    "squizlabs/php_codesniffer": "^3.1"
+    "phpunit/phpunit": "^7.2",
+    "squizlabs/php_codesniffer": "^3.3"
   },
   "autoload": {
     "psr-4": {
@@ -42,5 +40,7 @@
     ],
     "test:lint": "./vendor/bin/phpcs --extensions=php --standard=PSR2 ./src/",
     "test:unit": "./vendor/bin/phpunit --testsuite api --configuration phpunit.xml --coverage-text"
-  }
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true
 }
